requirements.txt
numpy
slangpy
pyimgui